+++ ng_trajectory/criterions/profile/profiler.py
@@ -243,18 +243,19 @@ def forward_pass(points: numpy.ndarray, v_bwd: List[float], v_max: List[float], 
     return v_fwd, a, t
 
 
-def profileCompute(points: numpy.ndarray, overlap: int = 0, fd: TextIO = None) -> Tuple[numpy.ndarray, numpy.ndarray, numpy.ndarray]:
+def profileCompute(points: numpy.ndarray, overlap: int = 0, fd: TextIO = None, lap_time: bool = False) -> Tuple[numpy.ndarray, numpy.ndarray, numpy.ndarray]:
     """Compute the speed profile using overlap.
 
     Arguments:
     points -- points of a trajectory with curvature, nx3 numpy.ndarray
     overlap -- size of trajectory overlap, int, default 0 (disabled)
+    lap_time -- when True, time is appended with lap_time, bool, default False
 
     Returns:
     v_fwd -- speed profile of the trajectory after forward step, [m.s^-1],
              nx1 numpy.ndarray
     a -- final acceleration profile of the trajectory, [m.s^-2], nx1 numpy.ndarray
-    t -- time of reaching the points of the trajectory, [s], nx1 numpy.ndarray
+    t -- time of reaching the points of the trajectory, [s], nx1 or (n+1)x1 numpy.ndarray
     """
 
     # Overlap points
@@ -276,6 +277,9 @@ def profileCompute(points: numpy.ndarray, overlap: int = 0, fd: TextIO = None) -
     _a = a[:, numpy.newaxis]
     _t = t[:, numpy.newaxis]
 
+    if lap_time:
+        __t = _t[-overlap]
+
     # Remove overlap and convert to numpy.ndarray
     if overlap > 0:
         _v = overlapRemove(_v, overlap)
@@ -283,6 +287,9 @@ def profileCompute(points: numpy.ndarray, overlap: int = 0, fd: TextIO = None) -
         _t = overlapRemove(_t, overlap)
 
         # Fix time array
+        if lap_time:
+            _t = numpy.vstack((_t, __t))
+
         _t = numpy.subtract(_t, _t[0])
 
         bwd = overlapRemove(bwd, overlap)
@@ -295,6 +302,9 @@ def profileCompute(points: numpy.ndarray, overlap: int = 0, fd: TextIO = None) -
             fd.write("Curvature: %s\n" % str(listFlatten(cur.tolist())))
             fd.write("Final speed: %s\n" % str(listFlatten(_v.tolist())))
             fd.flush()
+    elif lap_time:
+        _t = numpy.vstack((_t, _t[0]))
+        _t[0] = 0
 
     return _v, _a, _t